+++ src/Modules/Admin/Widgets/Forms/Traits/ActiveFormFieldsTrait.php
@@ -6,8 +6,6 @@ namespace Hirtz\Cms\Modules\Admin\Widgets\Forms\Traits;
 
 use Hirtz\Skeleton\Widgets\Forms\Fields\InputField;
 use Hirtz\Skeleton\Widgets\Forms\Fields\SelectField;
-use Hirtz\Skeleton\Widgets\Forms\Fields\TextareaField;
-use Hirtz\Skeleton\Widgets\Forms\Fields\TinyMceField;
 use Hirtz\Skeleton\Widgets\Forms\Fields\TypeSelectField;
 use Stringable;
 
@@ -29,25 +27,5 @@ trait ActiveFormFieldsTrait
     {
         return InputField::make()
             ->property('name');
-    }
-
-    protected function getContentField(): ?Stringable
-    {
-        if (!$this->model->contentType) {
-            return null;
-        }
-
-        return $this->model->contentType === 'html'
-            ? TinyMceField::make()
-                ->property('content')
-                ->validator($this->model->htmlValidator)
-            : TextareaField::make()
-                ->property('content');
-    }
-
-    protected function getLinkField(): ?Stringable
-    {
-        return InputField::make()
-            ->property('link');
     }
 }